Refactor, GitIgnore, Disabled Data and More Pages
Some pretty big refactoring! Took out the custom widgets and hid them in the checklist view. Gitignore has been messed with, although it isn't working for me. Data has some pushed buggy code so all of that is shut off. Finally, started work on multiple pages and getting a user flow set up.
